refactor world management to use floats

+++ src/main/java/me/chriss99/worldmanagement/Chunk.java
@@ -1,4 +1,4 @@
 package me.chriss99.worldmanagement;
 
-public record Chunk(int[][] data) {
+public record Chunk(float[][] data) {
 }
